Add "include" and "export" macros for modules
The include macro will find a ".roym" file for the given module and
add the types to the environment. For example:

    import "./structural"

    let f o = o.x + o.y
    f structural.obj

Will read the "structural.roym" and compile into:

    var structural = require("./structural");
    var f = function(o) {
        return o.x + o.y;
    }
    f(structural.obj)

On node.js, the export macro will compile into an assignment to the
"exports" global. For example:

    let obj = {x: 1, y: 2, t: "test"}
    export obj

Will compile into:

    var obj = {
        "x": 1,
        "y": 2,
        "t": "test"
    };
    exports["obj"] = obj;

Work needs to be done to only output exported names into module
descriptors and only leak exported names to the browser.