Remove `data` global
In the process, I made tags act like normal functions. To do so I used
a nice JavaScript trick:

    var C = function() {
        if(!(this instanceof C)) {
            return new C();
        }
    };

Which allows `C` to be called as a normal function and allow
`instanceof` checks to still work.

A caveat is that you now always have to give an argument to
constructors:

    data A = B | C
    let c = C ()